+++ user-service/src/main/java/com/plagod/service/RefundQueryService.java
@@ -5,11 +5,21 @@ import com.plagod.vo.entitlement.RefundVO;
 
 public interface RefundQueryService {
 
-    RefundPageResult pageOwnRefunds(Long tenantId, Long userId, long current, long size, String status);
+    RefundPageResult pageOwnRefunds(
+            Long tenantId,
+            Long userId,
+            Integer current,
+            Integer size,
+            String status);
 
     RefundVO getOwnRefund(Long tenantId, Long userId, String refundNo);
 
-    RefundPageResult pageForAdmin(Long tenantId, long current, long size, Long userId, String status);
+    RefundPageResult pageForAdmin(
+            Long tenantId,
+            Integer current,
+            Integer size,
+            Long userId,
+            String status);
 
     RefundVO getForAdmin(Long tenantId, String refundNo);
 }
